Use AuthorizationCodeResourceDetails instead of OAuth2ClientProperties
Using the native bean instead of a partial copy of it (which is
effectively what OAuth2ClientProperties was) makes it more flexible for
external configuration. Unfortunately it has a knock on effect of
rather a large number of other changes (since OAuth2ClientProperties
was used quite a lot). Hopefully it makes sense to do this now.

Users need to switch config from tokenUri->accessTokenUri and
authorizationUri->userAuthorizationUri.

Fixes gh-26