Use OAuth2RestTemplate to refresh access token if needed in Zuul
Just passing the access token on downstream is good enough for simple
use cases, but when the token has been acquired by an OAuth2RestTemplate
in the application (as in the SSO case) then we can re-use the
template to auto-refresh the token if that is allowed. If it isn't then
we now throw BadCredentialsException.

Fixes gh-52